+++ src/main/java/com/kilo52/common/struct/DataFrame.java
@@ -795,9 +795,9 @@ public interface DataFrame extends Cloneable, Serializable, Iterable<Column> {
 	 * @param col The index of the column to search 
 	 * @param regex The regular expression to search for
 	 * @return A sub-DataFrame containing all rows that match the given regular expression.
-	 * 		   <br>Returns Null if nothing in the column matches the given regular expression
+	 * 		   <br>Returns an empty DataFrame if nothing in the column matches the given regular expression
 	 */
-	public DataFrame findAll(int col, String regex);
+	public DataFrame filter(int col, String regex);
 	
 	/**
 	 * Computes and returns a {@link DataFrame} containing all rows that match the specified regular
@@ -808,7 +808,23 @@ public interface DataFrame extends Cloneable, Serializable, Iterable<Column> {
 	 * @param colName The name of the Column to search
 	 * @param regex The regular expression to search for
 	 * @return A sub-DataFrame containing all rows that match the given regular expression.
-	 * 		   <br>Returns Null if nothing in the column matches the given regular expression
+	 * 		   <br>Returns an empty DataFrame if nothing in the column matches the given regular expression
+	 */
+	public DataFrame filter(String colName, String regex);
+	
+	/**
+	 * @deprecated Renamed to {@link #filter(int, String)}.<br>
+	 * 			   This method has been replaced and will be removed in a future release
+	 * @param col The index of the column to search
+	 * @param regex The regular expression to search for
+	 */
+	public DataFrame findAll(int col, String regex);
+	
+	/**
+	 * @deprecated Renamed to {@link #filter(String, String)}.<br>
+	 * 			   This method has been replaced and will be removed in a future release
+	 * @param colName The name of the Column to search
+	 * @param regex The regular expression to search for
 	 */
 	public DataFrame findAll(String colName, String regex);