+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+#### 1.1.0
+* Renamed *findAll()* methods to *filter()* in the DataFrame API. All *findAll()* methods are now deprecated and will be removed in a future release
+* All *filter()* methods now return an empty DataFrame instead of null when there are no matches
+* Return type of public methods in CSVFileReader and CSVFileWriter changed to allow method chaining
+
 #### 1.0.0 
 * Final version for open source release